+++ src/main/java/com/blogapp/controller/PostController.java
@@ -3,8 +3,6 @@ package com.blogapp.controller;
 import com.blogapp.model.Comment;
 import com.blogapp.model.MyUserDetails;
 import com.blogapp.model.Post;
-import com.blogapp.model.Tag;
-import com.blogapp.repository.UserRepository;
 import com.blogapp.service.PostService;
 import com.blogapp.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +15,8 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
-import java.util.*;
+import java.util.Objects;
+import java.util.Optional;
 
 @Controller
 public class PostController {
@@ -26,34 +25,14 @@ public class PostController {
     private TagService tagService;
     @Autowired
     private PostService postService;
-    @Autowired
-    private UserRepository userRepository;
 
 
     @PostMapping("/savePost")
-    public String savePost(@ModelAttribute("post") Post post, String tagString, String publishType, @AuthenticationPrincipal MyUserDetails userDetails) {
-        List<String> tagList = new ArrayList<>(Arrays.asList(tagString.split(" ")));
-
-        for (String eachTag : tagList) {
-            Tag tag = tagService.getTagByName(eachTag) == null ? new Tag() : tagService.getTagByName(eachTag);
-
-            tag.setName(eachTag);
-            post.getTag().add(tag);
-            tag.getPost().add(post);
-        }
-
-        post.setUser(userRepository.findByEmail(userDetails.getEmail()).get());
-
-        switch (publishType) {
-            case "0":
-                post.setIsPublished(false);
-                post.setUpdatedAt(new Date());
-                break;
-            case "1":
-                post.setIsPublished(true);
-                post.setPublishedAt(new Date());
-                break;
-        }
+    public String savePost(@ModelAttribute("post") Post post, String tagString, String publishType,
+                           @AuthenticationPrincipal MyUserDetails userDetails) {
+        tagService.saveTags(post, tagString);
+        postService.updatePostDate(post, publishType);
+        postService.saveUser(post, userDetails);
 
         postService.savePosts(post);
         postService.createExcerpt(post.getId());
@@ -89,30 +68,26 @@ public class PostController {
 
     @PreAuthorize("hasAnyRole('ADMIN', 'AUTHOR')")
     @GetMapping("/updatePost/{id}")
-    public String updatePost(@PathVariable(value = "id") Integer id, Model model, Post post) {
+    public String updatePost(@PathVariable(value = "id") Integer id, Model model, @AuthenticationPrincipal MyUserDetails user) {
         Optional<Post> optional = postService.getPostById(id);
-        StringBuilder tagString = new StringBuilder();
 
-        if (optional.isPresent()) {
-            post = optional.get();
-        } else {
+        if (optional.isEmpty())
             throw new RuntimeException("Post not found with id::" + id);
-        }
+        else if(!postService.isAuthorized(id, user))
+            throw new RuntimeException("Unauthorized Access");
 
-        for (Tag tag : post.getTag()) {
-            tagString.append(tag.getName()).append(" ");
-        }
-
-
-        model.addAttribute("post", post);
-        model.addAttribute("tagString", tagString.toString());
+        model.addAttribute("post", optional.get());
+        model.addAttribute("tagString", tagService.getTagString(optional.get()));
 
         return "newPost";
     }
 
     @PreAuthorize("hasAnyRole('ADMIN', 'AUTHOR')")
     @GetMapping("/delete/{id}")
-    public String deletePost(@PathVariable(value = "id") Integer id) {
+    public String deletePost(@PathVariable(value = "id") Integer id, @AuthenticationPrincipal MyUserDetails user) {
+        if(!postService.isAuthorized(id, user))
+            throw new RuntimeException("Unauthorized Access");
+
         postService.deletePostsById(id);
 
         return "redirect:/";
